+++ plugins/woocommerce-admin/packages/components/src/stepper/check-icon.js
@@ -1,0 +1,22 @@
+/** @format */
+
+export default () => (
+    <svg
+        role="img"
+        aria-hidden="true"
+        focusable="false"
+        width="18"
+        height="18"
+        viewBox="0 0 18 18"
+        fill="none"
+        xmlns="http://www.w3.org/2000/svg"
+    >
+        <mask id="mask0" mask-type="alpha" maskUnits="userSpaceOnUse" x="2" y="3" width="14" height="12">
+            <path d="M6.59631 11.9062L3.46881 8.77875L2.40381 9.83625L6.59631 14.0287L15.5963
+                5.02875L14.5388 3.97125L6.59631 11.9062Z" fill="white" />
+        </mask>
+        <g mask="url(#mask0)">
+            <rect width="18" height="18" fill="white" />
+        </g>
+    </svg>
+);